requirements.txt
# Web框架
streamlit>=1.28.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 数据库
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# 可视化
plotly>=5.17.0

# 股票数据
akshare>=1.17.85

# 工具库
python-dotenv>=1.0.0
openpyxl>=3.0.0
pytz>=2023.3

